+++ app/models/his/his_hosxpv4.js
@@ -1214,6 +1214,7 @@ class HisHosxpv4Model {
         }
         let query = db({ o: "oapp" })
             .join({ c: "clinic" }, "c.clinic", "o.clinic")
+            .join({ ovst: "ovst" }, "o.vn", "ovst.vn")
             .join({ d: "doctor" }, "d.code", "o.doctor")
             .join({ oa: "oapp_status" }, "oa.oapp_status_id", "o.oapp_status_id")
             .leftJoin({ pt: "pttype" }, "o.next_pttype", "pt.pttype");
@@ -1228,10 +1229,10 @@ class HisHosxpv4Model {
             .select([
             "o.hn", "o.an", "o.vn", "o.visit_vn",
             db.raw("CASE WHEN o.patient_visit = 'Y' THEN 1 ELSE 0 END AS isvisited"),
-            db.raw("o.vstdate AS visit_date"),
-            db.raw("o.nextdate AS fu_date"),
-            db.raw("o.nexttime AS fu_time"),
-            db.raw("o.clinic AS cliniccode"),
+            db.raw("concat(o.vstdate,' ',ovst.vsttime) AS visit_date"),
+            db.raw("o.nextdate AS apdate"),
+            db.raw("o.nexttime AS aptime"),
+            "o.clinic",
             db.raw("c.name AS clinicName"),
             db.raw("o.doctor AS dr_code"),
             db.raw("d.name AS dr_name"),
@@ -1242,8 +1243,9 @@ class HisHosxpv4Model {
             db.raw("CASE WHEN o.note1 IS NULL OR o.note1 = '' THEN o.perform_text ELSE o.note1 END AS prepare_text"),
             db.raw("o.lab_list_text AS lab"),
             db.raw("o.xray_list_text AS xray"),
-            db.raw("o.contact_point AS visit_area"),
+            db.raw("o.contact_point AS apvisit_area"),
             db.raw("CASE WHEN o.oapp_status_id = 1 THEN 1 ELSE 0 END AS isactive"),
+            db.raw("o.update_datetime as d_update")
         ])
             .whereNotNull("o.nextdate")
             .whereRaw("o.vstdate < o.nextdate");