+++ src/models/his/his_hospitalos.ts
@@ -8,6 +8,11 @@ export class HisHospitalOsModel {
     constructor() {
     }
 
+    async tableExist(db: Knex, tableName: string): Promise<boolean> {
+        const exist = await db.schema.hasTable(tableName);
+        return exist;
+    }
+    
     // ❌ ไม่พบการเรียกใช้งาน
     check() {
         return true;